query builder separated in a util
