+++ src/user/user.service.ts
@@ -25,6 +25,7 @@ import * as crypto from 'crypto';
 import { ConfigService } from '@nestjs/config';
 import { FileUpload } from 'graphql-upload-minimal';
 import { Readable } from 'stream';
+import { UserQueryBuilderUtil } from 'src/common/user-query-builder.util';
 
 async function streamToBuffer(stream: Readable): Promise<Buffer> {
   return new Promise((resolve, reject) => {
@@ -58,36 +59,15 @@ export class UserService {
     sortBy?: string,
     sortOrder?: 'asc' | 'desc',
   ): Promise<PaginatedUserWithNumberOfPosts> {
-    // Construct order clause based on sortBy and sortOrder
-    const order: { [key: string]: 'ASC' | 'DESC' } = {};
-    if (sortBy) {
-      const sortField = sortBy === 'posts' ? 'postscount' : 'user.name';
-      order[sortField] = (sortOrder?.toUpperCase() as 'ASC' | 'DESC') || 'ASC';
-    }
-
-    const queryBuilder = this.userRepository
-      .createQueryBuilder('user')
-      .leftJoinAndSelect('user.posts', 'post') // Join with posts to count them
-      .leftJoin('user.role', 'role') // Join with role to access role properties
-      .select([
-        'user.id AS user_id',
-        'user.name AS user_name',
-        'user.email AS user_email',
-        'role.name AS user_role', // Accessing role name correctly
-        'COUNT(post.id) AS postscount', // Count posts and alias as postscount
-      ])
-      .groupBy('user.id') // Group by user ID
-      .addGroupBy('role.name') // Group by role name
-      .offset((page - 1) * limit) // Offset for pagination
-      .limit(limit); // Limit the results
-
-    if (role) {
-      queryBuilder.andWhere('role.name = :role', { role }); // Corrected to access role name
-    }
-
-    if (Object.keys(order).length) {
-      queryBuilder.orderBy(order);
-    }
+    const queryBuilder = this.userRepository.createQueryBuilder('user');
+    UserQueryBuilderUtil.buildUserWithPostsQuery(
+      queryBuilder,
+      page,
+      limit,
+      role,
+      sortBy,
+      sortOrder,
+    );
 
     // Fetch paginated users and total count
     const [users, countResult] = await Promise.all([
